Add User Roles Admin/Premium/Standard

+++ app/controllers/wikis_controller.rb
@@ -15,8 +15,8 @@ class WikisController < ApplicationController
   end
 
   def create
-    @wiki = Wiki.new
-    set_wiki_parameters
+    @wiki = Wiki.new(wiki_parameters)
+    set_new_wiki_user
 
     if @wiki.save
       flash[:notice] = wiki_instantiate_confirmed
@@ -29,13 +29,13 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-    set_wiki_parameters
 
-    if @wiki.save
+    if @wiki.update(wiki_parameters)
       flash[:notice] = wiki_update_confirmed
       redirect_to @wiki
     else
@@ -46,6 +46,7 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = wiki_destroy_confirmed
@@ -58,10 +59,12 @@ class WikisController < ApplicationController
 
   private
 
-  def set_wiki_parameters
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
-    @wiki.private = params[:wiki][:private]
+  def wiki_parameters
+    params.require(:wiki).permit(:title, :body, :private)
+  end
+
+  def set_new_wiki_user
+    @wiki.user = current_user
   end
 
   #Messages for our alerts
